requirements.txt
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6-sip==13.10.2
mysql-connector-python==8.2.0
pandas==2.2.0
matplotlib==3.10.3
reportlab==4.1.0
openpyxl==3.1.2
plotly==5.18.0
python-pptx==0.6.23
python-dotenv==1.0.0 